+++ src/pages/RegisterPage/components/PasswordInput/usePasswordValidation.ts
