+++ src/pages/RegisterPage/components/UsernameInput/useNameValidation.ts
